+++ src/views/default/view.php
@@ -27,8 +27,14 @@ foreach ($generator->templates as $name => $path) {
 
     <?php $form = ActiveForm::begin([
         'id' => "$id-generator",
-        'successCssClass' => '',
-        'fieldConfig' => ['class' => ActiveField::className()],
+        'successCssClass' => 'is-valid',
+        'errorCssClass' => 'is-invalid',
+        'validationStateOn' => ActiveForm::VALIDATION_STATE_ON_INPUT,
+        'fieldConfig' => [
+            'class' => ActiveField::className(),
+            'hintOptions' => ['tag' => 'small', 'class' => 'form-text text-muted'],
+            'errorOptions' => ['class' => 'invalid-feedback']
+        ],
     ]); ?>
         <div class="row">
             <div class="col-lg-8 col-md-10" id="form-fields">
@@ -36,11 +42,11 @@ foreach ($generator->templates as $name => $path) {
                     'generator' => $generator,
                     'form' => $form,
                 ]) ?>
-                <?= $form->field($generator, 'template')->sticky()
+                <?= $form->field($generator, 'template')
+                    ->sticky()
+                    ->hint('Please select which set of the templates should be used to generated the code.')
                     ->label('Code Template')
-                    ->dropDownList($templates)->hint('
-                        Please select which set of the templates should be used to generated the code.
-                ') ?>
+                    ->dropDownList($templates) ?>
                 <div class="form-group">
                     <?= Html::submitButton('Preview', ['name' => 'preview', 'class' => 'btn btn-primary']) ?>